Move DaemonThreadFactory to *.dsl package
