+++ src/main/java/com/lmax/disruptor/dsl/DaemonThreadFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lmax.disruptor;
+package com.lmax.disruptor.dsl;
 
 import java.util.concurrent.ThreadFactory;
