+++ src/perftest/java/com/lmax/disruptor/queue/PingPongQueueLatencyTest.java
@@ -26,7 +26,7 @@ import java.util.concurrent.LinkedBlockingQueue;
 
 import org.HdrHistogram.Histogram;
 
-import com.lmax.disruptor.DaemonThreadFactory;
+import com.lmax.disruptor.dsl.DaemonThreadFactory;
 
 /**
  * <pre>
