+++ src/perftest/java/com/lmax/disruptor/sequenced/ThreeToOneSequencedThroughputTest.java
@@ -31,7 +31,7 @@ import com.lmax.disruptor.SequenceBarrier;
 import com.lmax.disruptor.support.ValueAdditionEventHandler;
 import com.lmax.disruptor.support.ValueEvent;
 import com.lmax.disruptor.support.ValuePublisher;
-import com.lmax.disruptor.DaemonThreadFactory;
+import com.lmax.disruptor.dsl.DaemonThreadFactory;
 
 /**
  * <pre>
