+++ src/test/java/com/lmax/disruptor/RingBufferTest.java
@@ -34,6 +34,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import com.lmax.disruptor.dsl.DaemonThreadFactory;
 import org.junit.Test;
 
 import com.lmax.disruptor.support.StubEvent;
