+++ src/test/java/com/lmax/disruptor/example/DynamiclyAddHandler.java
@@ -3,7 +3,7 @@ package com.lmax.disruptor.example;
 import com.lmax.disruptor.*;
 import com.lmax.disruptor.dsl.Disruptor;
 import com.lmax.disruptor.support.StubEvent;
-import com.lmax.disruptor.DaemonThreadFactory;
+import com.lmax.disruptor.dsl.DaemonThreadFactory;
 
 import java.util.concurrent.CountDownLatch;
